+++ app/src/main/java/com/withpeace/withpeace/WithpeaceApp.kt
@@ -1,20 +1,26 @@
 package com.withpeace.withpeace
 
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
 import com.withpeace.withpeace.navigation.WithpeaceNavHost
 import kotlinx.coroutines.launch
 
 
 @Composable
 fun WithpeaceApp(
-    startDestination: String
+    startDestination: String,
+    navController: NavHostController = rememberNavController(),
 ) {
     val snackBarHostState = remember { SnackbarHostState() }
     val coroutineScope = rememberCoroutineScope()
@@ -22,14 +28,26 @@ fun WithpeaceApp(
         snackBarHostState.showSnackbar(message)
     }
 
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentDestination = navBackStackEntry?.destination
+
     Scaffold(
+        bottomBar = {
+            MainBottomBar(
+                currentDestination = currentDestination ?: return@Scaffold,
+                navController = navController,
+            )
+        },
         modifier = Modifier.fillMaxSize(),
         snackbarHost = { SnackbarHost(snackBarHostState) },
-    ) {
+    ) { innerPadding ->
         WithpeaceNavHost(
+            modifier = Modifier.padding(innerPadding),
             onShowSnackBar = ::showSnackBar,
-            startDestination = startDestination
+            startDestination = startDestination,
+            navController = navController,
         )
-        it
     }
+
 }
+
