+++ app/src/main/java/com/withpeace/withpeace/navigation/NavHost.kt
@@ -1,31 +1,38 @@
 package com.withpeace.withpeace.navigation
 
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navOptions
+import com.withpeace.withpeace.feature.home.navigation.homeNavGraph
 import com.withpeace.withpeace.feature.login.navigation.LOGIN_ROUTE
 import com.withpeace.withpeace.feature.login.navigation.loginNavGraph
-import com.withpeace.withpeace.mainNavGraph
+import com.withpeace.withpeace.feature.mypage.navigation.myPageNavGraph
+import com.withpeace.withpeace.feature.post.navigation.postNavGraph
 
 
 @Composable
 fun WithpeaceNavHost(
     modifier: Modifier = Modifier,
-    navController: NavHostController = rememberNavController(),
+    navController: NavHostController,
     startDestination: String = LOGIN_ROUTE,
     onShowSnackBar: (message: String) -> Unit,
 ) {
-    NavHost(
-        modifier = modifier,
-        navController = navController,
-        startDestination = startDestination,
-    ) {
-        loginNavGraph(onShowSnackBar = onShowSnackBar)
-        mainNavGraph(
-            onShowSnackBar = onShowSnackBar,
+    Box(modifier = modifier) {
+        NavHost(
+            modifier = modifier,
             navController = navController,
-        )
+            startDestination = startDestination,
+        ) {
+            loginNavGraph(onShowSnackBar = onShowSnackBar)
+            homeNavGraph(onShowSnackBar)
+            postNavGraph(onShowSnackBar)
+            myPageNavGraph(onShowSnackBar)
+        }
     }
+
 }
